TODO 해결: 네트워크 연결 실패 시

+++ app/src/main/java/amp/triples/Grid.kt
@@ -1,6 +1,6 @@
 package amp.triples
 
 data class Grid(
-    val xo: Int,
-    val yo: Int
+    val xo: Float,
+    val yo: Float
 )
